+++ src/objects.js
@@ -84,7 +84,7 @@ BlockEditorMorph, BlockDialogMorph, PrototypeHatBlockMorph,  BooleanSlotMorph,
 localize, TableMorph, TableFrameMorph, normalizeCanvas, VectorPaintEditorMorph,
 AlignmentMorph, Process, WorldMap, copyCanvas, useBlurredShadows*/
 
-modules.objects = '2021-February-20';
+modules.objects = '2021-February-23';
 
 var SpriteMorph;
 var StageMorph;
@@ -1409,12 +1409,14 @@ SpriteMorph.prototype.initBlocks = function () {
             spec: 'reshape %l to %nums',
             defaults: [null, [4, 3]]
         },
-        reportSlice: {
+    /*
+        reportSlice: { // currently not in use
             type: 'reporter',
             category: 'lists',
             spec: 'slice %l by %nums',
             defaults: [null, [2, -1]]
         },
+    */
 
         // HOFs
         reportMap: {
@@ -1761,10 +1763,6 @@ SpriteMorph.prototype.blockAlternatives = {
     doChangeVar: ['doSetVar'],
     doShowVar: ['doHideVar'],
     doHideVar: ['doShowVar'],
-
-    // lists:
-    reportReshape: ['reportSlice'],
-    reportSlice: ['reportReshape'],
 
     // HOFs
     reportMap: ['reportKeep', 'reportFindFirst'],